Refactor core operations in core_op.rs
